Eval loss is now always available as a metric
It is often handy to inspect the dependence of the pointwise loss on
values of a specific column. IIUC the only way of doing this with TFMA
prior to this commit is adding a "fake" metric

    eval_metric_ops={"eval-loss": tf.metrics.mean(loss)}

The name "loss" is reserved for the actual loss, therefore the users
are forced to use something different.