Support widget interoperability inside notebook.
When a row is selected in slicing metrics view in the notebook environment, fire an event with details about the slice being selected. This can be used for coordinating different widgets in the same Jupyter / Colab notebook.

The name of the event is "tfma-event" and its event detail will always have 2 fields:
- eventType stands for the type of the event.
- eventDetail contains the actual detail of the event.

The extra redicretion makes it easier to add other events in the future as there is no need to change the widget code for Jupyer / Colab. Implmenting it with mixin allows easy code reuse on other components if desired.

For example, the user can use TFMA to identify interesting slices and drill down furtehr with What-If Tool:
  https://pair-code.github.io/what-if-tool/

PiperOrigin-RevId: 248438091