Roll forward with fix for the ImportError when running on Cloud DataFlow.
Original description below:

Support widget interoperability inside notebook.

Add an event listener to the event "tfma-event" in the notebook js code, pass the event information across the python-js bridge and trigger the appropriate event handler on the python side.

Allow the user to register event handler for slicing metrics view by providing a dict as a parameter of the rendering API where the key is the name of the actual event and the value is the event handler funciton.

Currently, we only support "slice-selected" event and the handler will be triggered with a dict of 'sliceName' and 'sliceValue'.

To use it,
  tfma.view.render_slicing_metrics(results, slicing_spec=..., event_handlers={'slice-selected': handle_row_selected})

The function handle_row_selected will be called with a dict with two keys: sliceName and sliceValue when a slice is selcted either in the metrics table or slice overview widget.

PiperOrigin-RevId: 249539027