Split positive and negative count in prediction distribution and show them as separate data points.
Make the underlying chart a combo chart. The total prediciton count is still a bar chart while the positive and negative counts are scatter plots. All three values will have their own y scale while we show dual y axes for the positive and negative counts. The positive count is the main one (shows up on the left hand side) and the negative one is the secondary one (shows up on the right hand side).

The actual count is available in the tooltip on hover.

PiperOrigin-RevId: 257098845