Load metrics and plots serialized using the new structured arrays of keys and values and construct nested dictionary representing them.
For simplicity on visualization and backward compatibility, old data will also be loaded into the same nested structure with default key setting.

For metrics, the whole nested structure will be passed to the client side and the user can pick the configurations they want for visualization.

For plots, due to size consideration, the user will be required to specify a particular plot value to be visualized.

On the visualization side, add tfma-config-picker to enable the user to choose the desired set of metrics to see in slicing metrics view and time series view.

PiperOrigin-RevId: 267013058