Relax the duck typing check for confusion matrix at threashold.
As long as one of the confusion matrix value (TP, TN, FP, FN) is set in all elements of the matrices array, the duck typing check will pass. This will prevent the duck typing test from failing when some of the values are 0 (and omitted by protocol buffer).

Still check ALL fields are present when BoundedValue is used since it is a message type and will not be omitted by protocol buffer.

PiperOrigin-RevId: 270099880