Update merge/split_extracts implementation.
Adds the ability to control how (x,) and (x,1) arrays are split and merged. The previous implementation of split_extracts expanded the dimension of (x,) array instead of splitting it into scalars. To compensate, merge_extracts would flatten all arrays in shape (x,1). This makes it impossible to recreate original arrays in shape (x,1) as all original arrays of that shape would be flattened.

PiperOrigin-RevId: 469890448